src/app/services/openvoice_service.py
"""
Service for Text-to-Speech generation using OpenVoice (alternative to TTS).
OpenVoice is a newer instant voice cloning system that may have better transformers compatibility.
"""
import logging
import os
import sys
from pathlib import Path
from typing import Optional, List

# Add parent directories to path
sys.path.append(str(Path(__file__).parent.parent.parent))

logger = logging.getLogger(__name__)

# Try to import OpenVoice
OPENVOICE_AVAILABLE = False
try:
    # OpenVoice might be installed from GitHub
    import openvoice
    OPENVOICE_AVAILABLE = True
except ImportError:
    try:
        # Alternative import path
        from openvoice import se_extractor, openvoice_base
        OPENVOICE_AVAILABLE = True
    except ImportError:
        logger.warning("OpenVoice not available. Install with: pip install git+https://github.com/myshell-ai/OpenVoice.git")
        OPENVOICE_AVAILABLE = False

import torch
from pydub import AudioSegment

from app.config import DEFAULT_TTS_MODEL


class OpenVoiceService:
    """Service for generating audio using OpenVoice (alternative to TTS)."""
    
    def __init__(self, base_speakers_path: Optional[str] = None):
        """
        Initialize the OpenVoice service.
        
        Args:
            base_speakers_path: Path to OpenVoice base speakers directory
        """
        if not OPENVOICE_AVAILABLE:
            logger.error("OpenVoice is not available. Install from GitHub.")
            self.model = None
            self.device = "cpu"
            return
        
        self.device = "cuda" if torch.cuda.is_available() else "cpu"
        self.base_speakers_path = base_speakers_path
        self.model = None
        self._initialize_model()
    
    def _initialize_model(self):
        """Initialize the OpenVoice model."""
        if not OPENVOICE_AVAILABLE:
            self.model = None
            return
        
        try:
            logger.info("Initializing OpenVoice model...")
            # OpenVoice initialization code would go here
            # This is a placeholder - actual implementation depends on OpenVoice API
            logger.info("OpenVoice model initialized successfully")
        except Exception as e:
            logger.error(f"Failed to initialize OpenVoice model: {e}")
            self.model = None
    
    def generate_audio(
        self,
        text: str,
        voice_sample_path: str,
        output_path: Optional[str] = None,
        language: str = "en"
    ) -> str:
        """
        Generate audio from text using OpenVoice.
        
        Args:
            text: Text to convert to speech
            voice_sample_path: Path to the reference voice sample
            output_path: Optional output path for the audio file
            language: Language code (default: "en")
            
        Returns:
            Path to the generated audio file
        """
        if not self.model:
            raise RuntimeError("OpenVoice model not initialized")
        
        # Implementation would go here
        # This is a placeholder
        raise NotImplementedError("OpenVoice integration not yet implemented")
    
    def generate_audio_chunks(
        self,
        text: str,
        voice_sample_path: str,
        output_dir: Optional[str] = None,
        language: str = "en"
    ) -> List[str]:
        """
        Generate audio chunks from text.
        
        Args:
            text: Text to convert to speech
            voice_sample_path: Path to the reference voice sample
            output_dir: Optional output directory for chunk files
            language: Language code (default: "en")
            
        Returns:
            List of paths to generated audio chunk files
        """
        if not self.model:
            raise RuntimeError("OpenVoice model not initialized")
        
        # Implementation would go here
        raise NotImplementedError("OpenVoice integration not yet implemented")

