src/app/requirements.txt
# Reporter AI Web App Dependencies
# Note: echomimic_v3 has its own requirements.txt in src/third_party/echomimic_v3/
# This file ensures compatibility with echomimic_v3 and voice dependencies

# Core dependencies that must be compatible with echomimic_v3
# echomimic_v3 requires: moviepy==2.2.1, tensorflow==2.15.0, scikit-image, albumentations
# These packages require specific numpy and networkx versions

# NumPy: Must be >=1.24.4 for albumentations/scikit-image, >=1.25.0 for moviepy, <2.0.0 for tensorflow
numpy>=1.25.0,<2.0.0

# NetworkX: Must be >=3.0 for scikit-image 0.25.2
networkx>=3.0

# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# HTTP client for article fetching and image downloads
aiohttp>=3.9.0
requests>=2.31.0

# HTML parsing
beautifulsoup4>=4.12.0

# LLM API clients
anthropic>=0.18.0
openai>=1.0.0

# Environment variables
python-dotenv>=1.0.0

# YAML parsing (for character configs)
pyyaml>=6.0

# Image processing
Pillow>=9.0.0

# TTS Service Dependencies
# Coqui TTS library for text-to-speech
# Using your_tts model (configured in config.py) which works with transformers>=4.46.2
# Compatibility patch (tts_compat.py) handles BeamSearchScorer import
TTS>=0.22.0

# PyTorch (required for TTS)
# Note: echomimic_v3 requires torch>=2.1.2, so we align with that
# Install based on your CUDA version if using GPU
# For CPU: pip install torch torchaudio torchvision --index-url https://download.pytorch.org/whl/cpu
# For CUDA 11.8: pip install torch torchaudio torchvision --index-url https://download.pytorch.org/whl/cu118
# For CUDA 12.1: pip install torch torchaudio torchvision --index-url https://download.pytorch.org/whl/cu121
torch>=2.1.2
torchaudio>=2.0.0
torchvision>=0.15.0

# Transformers (required for TTS models and echomimic_v3)
# TTS 0.22.0 requires transformers>=4.33.0
# echomimic_v3 requires transformers>=4.46.2
# So we need transformers>=4.46.2 to satisfy both
# BeamSearchScorer is available in transformers.generation.beam_search (compatibility patch handles this)
transformers>=4.46.2

# HuggingFace Hub (required for downloading TTS models)
huggingface-hub>=0.20.0,<1.0

# Audio processing (for TTS audio chunking and combining)
pydub>=0.25.1

# Qwen VLM (optional - for local model support)
# Uncomment if using Qwen VLM
# qwen-vl-utils>=0.1.0
# bitsandbytes>=0.41.0
# accelerate>=0.24.0

